feat: move to a single-export package pattern for libs
Several excellent points have been brought up related to the complexity
of how our libraries are built/deployed/published. A move towards a
single-export package structure simplifies several aspects of our
libraries:

It forces developers to use the entry-point in the `package.json` file,
and to use named imports for the components.

- This allows Jest to use CJS for tests

- This allows Webpack to use strict ESM

- This allows us to publish from the root package and deprecate the
  `publish-build` script

- Tree-shaking will work since Webpack resolves to the `module` field
  and will use the ESM build (verified in CRA2.0)

- It will be easier to search through our polyrepo codebase for usages
  as the named export must be referenced as is defined

BREAKING CHANGE: All imports must move to named imports from the
entry-point.
