+++ docs/.env
@@ -1,1 +1,1 @@
-NODE_PATH=../build
+#NODE_PATH=../build
