+++ docs/src/demos/Organisms/index.js
@@ -1,7 +1,5 @@
 import React from 'react'
 
-export * from './HeaderBar'
-
 export class Organisms extends React.Component {
     static id = 'organisms'
