+++ docs/src/helpers.js
@@ -1,5 +1,6 @@
 import React from 'react'
-import { Divider as CoreDivider } from 'core/Divider'
+
+import { Divider as CoreDivider } from '@dhis2/ui-core'
 
 export const Title = ({ children }) => <h1 className="title">{children}</h1>
 export const Heading = ({ children }) => <h2 className="heading">{children}</h2>
