+++ src/CircularProgress/styles.js
