+++ src/LinearProgress/index.js
