+++ src/LinearProgress/styles.js
