+++ src/Switch/index.js
