+++ src/Switch/styles.js
