+++ src/index.js
@@ -1,20 +1,21 @@
-// Core
-export { default as Button } from './core/Button'
-export { default as Card } from './core/Card'
-export { default as Checkbox } from './core/Checkbox'
-export { default as Chip } from './core/Chip'
-export { default as CircularProgress } from './core/CircularProgress'
-export { default as Divider } from './core/Divider'
-export { default as DropdownButton } from './core/DropdownButton'
-export { default as Help } from './core/Help'
-export { default as InputField } from './core/InputField'
-export { default as LinearProgress } from './core/LinearProgress'
-export { default as Logo } from './core/Logo'
-export { default as Menu } from './core/Menu'
-export { default as Radio } from './core/Radio'
-export { default as SelectField } from './core/SelectField'
-export { default as SplitButton } from './core/SplitButton'
-export { default as Switch } from './core/Switch'
+export { default as Button } from './Button'
+export { default as Card } from './Card'
+export { default as Checkbox } from './Checkbox'
+export { default as Chip } from './Chip'
+export { default as CircularProgress } from './CircularProgress'
+export { default as Divider } from './Divider'
+export { default as DropdownButton } from './DropdownButton'
+export { default as Help } from './Help'
+export { default as InputField } from './InputField'
+export { default as LinearProgress } from './LinearProgress'
 
-// Widgets
-export { default as HeaderBar } from './widgets/HeaderBar'
+export { default as Logo } from './Logo'
+export { LogoIcon, LogoIconWhite, LogoWhite } from './Logo'
+
+export { default as Menu } from './Menu'
+export { default as Radio } from './Radio'
+export { default as SelectField } from './SelectField'
+export { default as SplitButton } from './SplitButton'
+export { default as Switch } from './Switch'
+
+export { theme, colors, fonts } from './theme.js'
