+++ src/utils/index.js
@@ -4,5 +4,4 @@ export * from './css'
 export * from './rtl'
 export * from './math'
 export * from './timings'
-export * from './api'
 export * from './animations/index'
