--- src/widgets/HeaderBar/index.js
@@ -1,7 +1,0 @@
-import { HeaderBar } from './HeaderBar.js'
-export { HeaderBar }
-
-import { HeaderBarContainer } from './HeaderBar.Container.js'
-export { HeaderBarContainer }
-
-export default HeaderBarContainer
